Revert "feat: rework trace (#1062)"
This reverts commit 47c4292b367a7c1dc920070eff194642a6b8db7b.
