+++ examples/dev/metadata.eg.ts
@@ -8,7 +8,6 @@
  * even plain-text descriptions. This metadata serves as the basis off of which we
  * generate chain-specific APIs. Unless you are building an advanced Capi-based library,
  * chances are that you don't need to work with the metadata directly.
- * @test_skip
  */
 
 import { polkadotDev } from "@capi/polkadot-dev"
